requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloudinary==1.32.0
defusedxml==0.7.1
dj-database-url==2.0.0
dj3-cloudinary-storage==0.0.6
Django==4.1.7
django-paypal==2.0
fonttools==4.39.3
fpdf2==2.7.3
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.7
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2022.7
urllib3==1.26.15
whitenoise==6.5.0